requirements.txt
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
pyyaml>=6.0
pytest>=7.0.0
pytest-cov>=4.0.0
seaborn>=0.12.0
pandas>=1.5.0
tqdm>=4.65.0